Added the concept of Species and Group
Species is the NFTs metadata identity, basically like what would all be the same NFT id if this were an ERC1155.

Group is an arbitrary typing that all members of this species belong to. This could be used to classify things as being Items/Champs/Skins etc.

SBTs can now be unlocked en-masse by species and by group.